+++ source/const.py
@@ -3,9 +3,9 @@ RESOLUTION = (1280, 720)
 PERSO_H_SCALE = 0.5
 PERSO_V_SCALE = 0.8
 
-# Map
-MAP_HEIGHT = 10
+# Map (taille mini = 5x5)
 MAP_WIDTH = 10
+MAP_HEIGHT = 10
 
 # Controles
 UP=(0,-1)
